+++ src/main/java/org/example/remedy/application/auth/AuthServiceImpl.java
@@ -3,6 +3,7 @@ package org.example.remedy.application.auth;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.example.remedy.application.auth.port.in.AuthService;
 import org.example.remedy.presentation.auth.dto.AuthLoginRequest;
 import org.example.remedy.presentation.auth.dto.AuthRegisterRequest;
 import org.example.remedy.application.auth.exception.InvalidPasswordException;
